Add option for specific seed file for Novoplasty

+++ src/python/jobs/NovoplastyJob.py
@@ -24,6 +24,7 @@ class NovoplastyJob(Job):
         config_file_id,
         config_file_name,
         chained_job=False,
+        seed_file_id=False,
         parent_rv={},
         *args,
         **kwargs
@@ -51,6 +52,7 @@ class NovoplastyJob(Job):
         self.config_file_name = config_file_name
         self.chained_job = chained_job
         self.parent_rv = parent_rv
+        self.seed_file_id = seed_file_id
 
     def run(self, fileStore):
         """
@@ -95,16 +97,20 @@ class NovoplastyJob(Job):
                 read_two_file_path = utilities.readGlobalFile(
                     fileStore, self.read_two_file_id, bbnorm_params["read_two_file_name"]
                 )
-                with open(seed_file_path, "w+") as f:
-                    with open(read_one_file_path, "rt") as g:
-                        do_write = False
-                        for line in g:
-                            if line[0] == "@":
-                                do_write = True
-                            if line[0] == "+":
-                                break
-                            if do_write:
-                                f.write(line)
+
+                #Check if seed file exists; I.E., it's been imported
+                if not os.path.isfile(seed_file_path):
+                    # Create seed file from first read
+                    with open(seed_file_path, "w+") as f:
+                        with open(read_one_file_path, "rt") as g:
+                            do_write = False
+                            for line in g:
+                                if line[0] == "@":
+                                    do_write = True
+                                if line[0] == "+":
+                                    break
+                                if do_write:
+                                    f.write(line)
 
             else:
                 # Read the read files from the file store into the local
@@ -115,16 +121,20 @@ class NovoplastyJob(Job):
                 read_two_file_path = utilities.readGlobalFile(
                     fileStore, self.read_two_file_id, common_config["read_two_file_name"]
                 )
-                with open(seed_file_path, "w+") as f:
-                    with gzip.open(read_one_file_path, "rt") as g:
-                        do_write = False
-                        for line in g:
-                            if line[0] == "@":
-                                do_write = True
-                            if line[0] == "+":
-                                break
-                            if do_write:
-                                f.write(line)
+
+                #TODO: Check if seed file exists; I.E., it's been imported
+                if not os.path.isfile(seed_file_path):
+                    # Create seed file from first read
+                    with open(seed_file_path, "w+") as f:
+                        with gzip.open(read_one_file_path, "rt") as g:
+                            do_write = False
+                            for line in g:
+                                if line[0] == "@":
+                                    do_write = True
+                                if line[0] == "+":
+                                    break
+                                if do_write:
+                                    f.write(line)
 
             # Write the NOVOPlasty config file into the local temporary
             # directory
@@ -267,6 +277,16 @@ if __name__ == "__main__":
         help="path to a .ini file with args to be passed to the assembler",
     )
     parser.add_argument(
+        "--seed-path",
+        default=os.sep + os.path.join(*cmps),
+        help="path to a .fastq file to be used as seed by assembler, where applicable (defaults to using first read)",
+    )
+    parser.add_argument(
+        "--seed-file",
+        default=None,
+        help="path to a .fastq file to be used as seed by assembler, where applicable (defaults to using first read)",
+    )
+    parser.add_argument(
         "-o",
         "--output-directory",
         default=None,
@@ -295,6 +315,14 @@ if __name__ == "__main__":
             config_file_id = utilities.importConfigFile(
                 toil, os.path.join(options.config_path, options.config_file)
             )
+
+            # Import local seed file into the file store, if needed
+            if options.seed_file:
+                seed_file_id = utilities.importFile(
+                    toil, os.path.join(options.seed_path, options.seed_file)
+                )
+            else:
+                seed_file_id = None
 
             # Construct and start the NOVOPlasty job
             novoplasty_job = NovoplastyJob(
